+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/Verifier.kt
@@ -118,4 +118,4 @@ fun CryptoPublicKey.matchesIdentifier(input: String): Boolean {
     return false
 }
 
-class VerificationError(message: String?) : Throwable(message)
+class VerificationError(cause: Throwable) : Throwable(cause)