+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/validation/sdJwt/SdJwtCredentialPayloadValidationSummary.kt
@@ -1,12 +1,10 @@
 package at.asitplus.wallet.lib.agent.validation.sdJwt
 
-import at.asitplus.wallet.lib.agent.validation.common.SubjectMatchingResult
 import at.asitplus.wallet.lib.data.VerifiableCredentialSdJwt
 
 data class SdJwtCredentialPayloadValidationSummary(
     val verifiableCredentialSdJwt: VerifiableCredentialSdJwt,
-    val subjectMatchingResult: SubjectMatchingResult?,
 ) {
     val isSuccess: Boolean
-        get() = subjectMatchingResult?.isSuccess != false
+        get() = true
 }